rename: infrastructure 모듈로 이동
[#64]
