+++ build.gradle.kts
@@ -69,10 +69,3 @@ project(":batch") {
         implementation(project(":domain:mongo"))
     }
 }
-
-project(":command") {
-    dependencies {
-        implementation(project(":common"))
-        implementation(project(":domain:mysql"))
-    }
-}
