+++ command/build.gradle.kts
@@ -3,30 +3,14 @@ plugins {
 
     kotlin("jvm")
     kotlin("plugin.spring")
-    kotlin("plugin.allopen")
-    kotlin("plugin.jpa")
-    kotlin("kapt")
-}
-
-allOpen {
-    annotation("javax.persistence.Entity")
-    annotation("javax.persistence.MappedSuperclass")
-    annotation("javax.persistence.Embeddable")
 }
 
 dependencies {
+    implementation(project(":common"))
+    implementation(project(":domain:mysql"))
+    implementation(project(":infrastructure:rdbms:jpa"))
+
     implementation("org.springframework.boot:spring-boot-starter-amqp")
-    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     implementation("org.springframework.boot:spring-boot-starter-validation")
     implementation("org.springframework.boot:spring-boot-starter-web")
-    implementation("com.querydsl:querydsl-jpa:4.4.0")
-    kapt("com.querydsl:querydsl-apt:4.4.0:jpa")
-    annotationProcessor("org.springframework.boot:spring-boot-configuration-processor")
-    annotationProcessor(group = "com.querydsl", name = "querydsl-apt", classifier = "jpa")
-    runtimeOnly("mysql:mysql-connector-java")
-    runtimeOnly("com.h2database:h2")
-}
-
-sourceSets["main"].withConvention(org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet::class) {
-    kotlin.srcDir("$buildDir/generated/source/kapt/main")
 }
