--- command/src/main/kotlin/com/hs/infrastructure/jpa/repository/ProductEventLogSpringDataJpaRepository.kt
@@ -1,6 +1,0 @@
-package com.hs.infrastructure.jpa.repository
-
-import com.hs.infrastructure.jpa.persistence.ProductEventLogPersistence
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface ProductEventLogSpringDataJpaRepository : JpaRepository<ProductEventLogPersistence, Long>
