+++ domain/mysql/build.gradle.kts
@@ -1,14 +1,4 @@
-import org.springframework.boot.gradle.tasks.bundling.BootJar
-
-val jar: Jar by tasks
-val bootJar: BootJar by tasks
-
-bootJar.enabled = false
-jar.enabled = true
-
 plugins {
-    id("org.springframework.boot")
-
     kotlin("jvm")
 }
