+++ domain/mysql/src/main/kotlin/com/hs/repository/ProductEventLogRepository.kt
@@ -1,0 +1,7 @@
+package com.hs.repository
+
+import com.hs.entity.Product
+
+interface ProductEventLogRepository {
+    fun save(productId: Long, eventStatus: Product.EventStatus, message: String)
+}
