+++ infrastructure/rdbms/jpa/src/main/kotlin/com/hs/config/property/WriteProperty.kt
