+++ infrastructure/rdbms/jpa/src/main/kotlin/com/hs/entity/ProductImagePersistence.kt
@@ -1,4 +1,4 @@
-package com.hs.infrastructure.jpa.persistence
+package com.hs.entity
 
 import org.hibernate.annotations.DynamicUpdate
 import javax.persistence.Column
