+++ infrastructure/rdbms/jpa/src/main/kotlin/com/hs/repository/ProductQuerydslRepository.kt
@@ -1,17 +1,16 @@
-package com.hs.infrastructure.querydsl
+package com.hs.repository
 
 import com.hs.entity.Product
-import com.hs.infrastructure.jpa.persistence.ProductPersistence
-import com.hs.infrastructure.jpa.persistence.QProductImagePersistence.productImagePersistence
-import com.hs.infrastructure.jpa.persistence.QProductPersistence.productPersistence
-import com.hs.repository.ProductQueryRepository
+import com.hs.entity.QProductImagePersistence.productImagePersistence
+import com.hs.entity.QProductPersistence.productPersistence
+import com.hs.entity.ProductPersistence
 import com.querydsl.jpa.impl.JPAQueryFactory
 import org.springframework.stereotype.Repository
 import org.springframework.transaction.annotation.Transactional
 
 @Repository
 @Transactional(readOnly = true)
-class ProductQueryRepositoryImpl(private val queryFactory: JPAQueryFactory) : ProductQueryRepository {
+class ProductQuerydslRepository(private val queryFactory: JPAQueryFactory) : ProductQueryRepository {
 
     override fun findProductWithFetchJoin(id: Long): Product? {
         val productPersistence: ProductPersistence? = queryFactory
